backend/myenv/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
bidict==0.23.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.110.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
joblib==1.4.0
keyboard==0.13.5
lightgbm==4.3.0
multidict==6.0.5
numpy==1.24.4
openai==0.28.0
pydantic==2.6.4
pydantic-core==2.16.3
pyserial==3.5
python-engineio==4.9.0
python-socketio==5.11.1
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
simple-websocket==1.0.0
sniffio==1.3.1
starlette==0.36.3
threadpoolctl==3.4.0
tqdm==4.66.2
typing-extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
wsproto==1.2.0
yarl==1.9.4
